backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.8
cloudinary==1.42.2
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.1.3
django-admin-interface==0.29.4
django-cloudinary-storage==0.3.0
django-colorfield==0.12.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
flatbuffers==25.2.10
frozenlist==1.5.0
gast==0.6.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
google-genai==1.5.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.1.0
namex==0.0.8
numpy==2.1.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.0.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
service-identity==24.2.0
setuptools==76.0.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.19.0
termcolor==2.5.0
text-unidecode==1.3
twilio==9.3.7
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zope.interface==7.2
